ACC-2052 rename CompositeAttribute to CompositAttributeImpl + add CompositeAttribute interface
CompositeAttribute contains getAttributeByName() and getAttributes() methods. Its implementations are CompositeAttributeImpl, ContentAttribute and UserAttribute.
