+++ contentgrid-appserver-application-model/src/main/java/com/contentgrid/appserver/application/model/attributes/CompositeAttribute.java
@@ -1,65 +1,18 @@
 package com.contentgrid.appserver.application.model.attributes;
 
-import com.contentgrid.appserver.application.model.attributes.flags.AttributeFlag;
-import com.contentgrid.appserver.application.model.exceptions.DuplicateElementException;
 import com.contentgrid.appserver.application.model.values.AttributeName;
 import com.contentgrid.appserver.application.model.values.ColumnName;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.Singular;
-import lombok.Value;
 
-@Value
-public class CompositeAttribute implements Attribute {
-
-    @NonNull
-    AttributeName name;
-
-    String description;
-
-    Set<AttributeFlag> flags;
-
-    @Getter(AccessLevel.NONE)
-    Map<AttributeName, Attribute> attributes = new HashMap<>();
-
-    @Builder
-    CompositeAttribute(@NonNull AttributeName name, String description, @Singular Set<Attribute> attributes, @Singular Set<AttributeFlag> flags) {
-        this.name = name;
-        this.description = description;
-        this.flags = flags;
-        for (var attribute : attributes) {
-            if (this.attributes.put(attribute.getName(), attribute) != null) {
-                throw new DuplicateElementException("Duplicate attribute named %s".formatted(attribute.getName()));
-            }
-        }
-        for (var flag : this.flags) {
-            flag.checkSupported(this);
-        }
-    }
+public sealed interface CompositeAttribute extends Attribute permits CompositeAttributeImpl, ContentAttribute, UserAttribute {
 
     /**
      * Returns an unmodifiable list of attributes.
      *
      * @return an unmodifiable list of attributes
      */
-    public List<Attribute> getAttributes() {
-        return List.copyOf(attributes.values());
-    }
-
-    @Override
-    public List<ColumnName> getColumns() {
-        return attributes.values().stream()
-                .map(Attribute::getColumns)
-                .flatMap(List::stream)
-                .toList();
-    }
+    List<Attribute> getAttributes();
 
     /**
      * Finds an Attribute by its name.
@@ -67,7 +20,16 @@ public class CompositeAttribute implements Attribute {
      * @param attributeName the name of the attribute to find
      * @return an Optional containing the Attribute if found, or empty if not found
      */
-    public Optional<Attribute> getAttributeByName(AttributeName attributeName) {
-        return Optional.ofNullable(attributes.get(attributeName));
+    default Optional<Attribute> getAttributeByName(AttributeName attributeName) {
+        return getAttributes().stream()
+                .filter(attribute -> attribute.getName().equals(attributeName))
+                .findAny();
+    }
+
+    @Override
+    default List<ColumnName> getColumns() {
+        return getAttributes().stream()
+                .flatMap(attribute -> attribute.getColumns().stream())
+                .toList();
     }
 }
